GRAD_CAM/requirements.txt
absl-py==0.13.0
asgiref==3.3.1
attrs==21.2.0
beautifulsoup4==4.9.3
boltons==20.2.1
category-encoders==2.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.21.0
cycler==0.10.0
Django==3.1.6
face==20.1.1
Flask==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonify==0.5
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.2
mediapipe==0.8.6
numpy==1.19.3
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
pandas==1.2.3
patsy==0.5.1
Pillow==8.0.1
protobuf==3.17.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
schedule==1.1.0
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
statsmodels==0.12.2
threadpoolctl==2.1.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
